requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
alembic
aiomysql
greenlet
python-jose[cryptography]
passlib[bcrypt]
slowapi
python-multipart
email-validator